Add a second list of status codes, for indicating "changed" result
During URI requests we handle status_code to see if the Task failed or
not. But we also need to distinguish changed or not.

Currently the module always returns ok-green, but REST API may be so
rich to return the status depending whether something was effected.

For example Graylog returns 400 when a user was already present and 201
when it was created.

I propose adding the `changed_code` option to the module to augment
status_code.